requirements_neural.txt
torch
torchvision
Pillow
onnx
onnx-simplifier